Eliminated most warnings (to my knowledge, the only warning left is related to strnlen). And, of course, the warnings because some of the error codes don't have any comments (from errorParser).
git-svn-id: https://vdsf.svn.sourceforge.net/svnroot/vdsf/trunk@149 53c26a9f-4021-0410-9f73-c4284c400f72

+++ src/Common/Tests/ProcessLock/LockConcurrency.c
@@ -181,7 +181,7 @@ int main( int argc, char* argv[] )
          ERROR_EXIT( expectedToPass, NULL, ; );
       
       sprintf( strNumChilds, "%d", numChilds );
-      sprintf( strTime, "%d", maxTime );
+      sprintf( strTime, "%u", (unsigned int)maxTime );
       if ( tryMode )
          strcpy( strMode, "try" );
       else
